refactor(topic): move enum classes - ImportanceLevel & ResourceType
enum 을 따로 모아두면 보기 편하기 때문에 리팩토링 했습니다

#6
