+++ src/main/java/com/kllhy/roadmap/roadmap/persistence/model/ResourceSubTopic.java
@@ -1,5 +1,6 @@
 package com.kllhy.roadmap.roadmap.persistence.model;
 
+import com.kllhy.roadmap.roadmap.persistence.model.enums.ResourceType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
