+++ src/main/java/com/kllhy/roadmap/roadmap/persistence/model/ResourceTopic.java
@@ -1,5 +1,6 @@
 package com.kllhy.roadmap.roadmap.persistence.model;
 
+import com.kllhy.roadmap.roadmap.persistence.model.enums.ResourceType;
 import jakarta.persistence.*;
 
 @Entity
