+++ src/main/java/com/kllhy/roadmap/roadmap/persistence/model/enums/ResourceType.java
@@ -1,4 +1,4 @@
-package com.kllhy.roadmap.roadmap.persistence.model;
+package com.kllhy.roadmap.roadmap.persistence.model.enums;
 
 public enum ResourceType {
     VIDEO,
